+++ src/components/StartActive.css
@@ -1,0 +1,44 @@
+.FSE-StartActive {
+    position: relative;
+}
+.FSE-StartActive.is-active {
+    color: #0078ff;
+}
+.FSE-StartActive .FSE-StartActive-progress {
+    display: none;
+}
+.FSE-StartActive.is-active .FSE-StartActive-progress {
+    background: #0056b7;
+    display: block;
+    height: 100%;
+    left: 0;
+    position: absolute;
+    top: 0;
+    transition: width .3s;
+    width: 0%;
+    z-index: 2;
+}
+.FSE-StartActive.is-active:before {
+    background:  #0078ff;
+    content: '';
+    display: block;
+    height: 100%;
+    left: 0;
+    position: absolute;
+    top: 0;
+    width: 100%;
+    z-index: 1;
+}
+.FSE-StartActive .FSE-StartActive-progressLabel {
+    display: none;
+}
+.FSE-StartActive.is-active .FSE-StartActive-progressLabel {
+    color: #fff;
+    display: block;
+    left: 0;
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 100%;
+    z-index: 3;
+}
