+++ src/main.css
@@ -1,0 +1,3 @@
+#FSE-root {
+    display: inline-block;
+}
